requirements.txt
altgraph==0.17.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cryptography==46.0.6
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
msal==1.35.1
packaging==26.1
pillow==12.2.0
platformdirs==4.9.6
pycparser==3.0
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.4
PyJWT==2.12.1
pyperclip==1.11.0
requests==2.33.1
smmap==5.0.3
urllib3==2.6.3
